Cleanup and Changes
- Updated networking documentation
- Build now sends the current git hash to the compiler
- Added point in world.proto
- Switched to floats for Point
- Cleanup unused code in RemoteClient::tryParse
- Cleanup server startup log line
- Cleanup logs, switch to info in some places
- Missed tracking world in last changelog, some changes from that

+++ src/geo.cpp
@@ -2,7 +2,7 @@
 
 #include "geo.h"
 
-Point::Point(double x, double y, double z) {
+Point::Point(float x, float y, float z) {
     this->x = x;
     this->y = y;
     this->z = z;
